+++ mako/cogs/orm.py
@@ -21,7 +21,7 @@ import discord
 from discord.ext import commands
 from tortoise import Tortoise
 
-from mako.db import User, Guild
+from mako.db import Guild, User
 
 
 class Orm(commands.Cog):
@@ -30,27 +30,38 @@ class Orm(commands.Cog):
 
     async def test(self):
         for current_guild in self.bot.guilds:
-            if not await Guild.filter(guild_id=current_guild.id).exists():
-                if not await User.filter(user_snowflake=current_guild.owner.id).exists():
-                    owner = await User.get_or_create(user_snowflake=current_guild.owner.id)
-                    guild = await Guild.get_or_create(guild_id=current_guild.id,
-                                                      notification_channel=current_guild.system_channel.id,
-                                                      owner=owner)
+            if not current_guild.system_channel:
+                notification_channel = 0
+            else:
+                notification_channel = current_guild.system_channel.id
+            if not await Guild.filter(id=current_guild.id).exists():
+                if not await User.filter(id=current_guild.owner.id).exists():
+                    owner, _ = await User.get_or_create(id=current_guild.owner.id)
+                    guild, _ = await Guild.get_or_create(id=current_guild.id,
+                                                                notification_channel=notification_channel,
+                                                                owner=owner)
                 else:
-                    owner = await User.get_or_create(user_snowflake=current_guild.owner.id)
-                    guild = await Guild.get_or_create(guild_id=current_guild.id,
-                                                      notification_channel=current_guild.system_channel.id,
-                                                      owner=owner)
-                print(f'Saved {guild[0].id}/{current_guild.name} with owner '
-                      f'{owner[0].id}/{str(current_guild.owner)} to Database.')
+                    owner, _ = await User.get_or_create(id=current_guild.owner.id)
+                    guild, _ = await Guild.get_or_create(id=current_guild.id,
+                                                                notification_channel=notification_channel,
+                                                                owner=owner)
+                print(f'Saved {guild.id} / {current_guild.name} with owner '
+                      f'{owner.id} / {str(current_guild.owner)} to Database.')
             else:
                 continue
         await Tortoise.close_connections()
+
+    async def get_guild_owner(self, guild_id):
+        guild, _ = await Guild.get()
 
     @discord.slash_command(guild_only=True, guild_ids=[656899959035133972, 1054741800671252532])
     async def test_commands(self, ctx) -> None:
         await self.test()
         await ctx.respond('Done')
+
+    @discord.slash_command(guild_only=True, guild_ids=[656899959035133972, 1054741800671252532])
+    async def get_owner(self, ctx) -> None:
+        await ctx.respond()
 
 
 def setup(bot):
