+++ requirements.txt
@@ -1,5 +1,4 @@
 py-cord==2.4.0
 requests==2.28.2
-aiosqlite==0.18.0
 twitchAPI==3.8.0
 tortoise-orm==0.19.3
